undo changes to movisens test and instead skip test if in CI environment as it depends on ability to download

+++ tests/testthat/test_greadaccfile.R
@@ -157,40 +157,59 @@ test_that("g.readaccfile and g.inspectfile can read movisens, gt3x, cwa, Axivity
                   inspectfileobject = IGA)
   expect_true(MGA$filetooshort)
   
-  cat("\n Movisens")
-  
-  output_dir = "output_unisensExample"
-  on.exit({if (file.exists(output_dir)) unlink(output_dir, recursive = TRUE)}, add = TRUE)
-  if (file.exists(output_dir)) unlink(output_dir, recursive = TRUE)
- 
-  movisensFile  = system.file("testfiles/unisensExample.bin", package = "GGIR")[1]
-  Mcsv = g.inspectfile(movisensFile, desiredtz = desiredtz)
-  expect_equal(Mcsv$monc, MONITOR$MOVISENS)
-  expect_equal(Mcsv$dformc, FORMAT$BIN)
-  expect_equal(Mcsv$sf, 64)
-  
-  movisens_blocksize = 3000
-  movisens_read = g.readaccfile(movisensFile, blocksize = movisens_blocksize, blocknumber = 1, filequality = filequality,
-                                dayborder = dayborder, ws = 3,
-                                PreviousEndPage = 1, inspectfileobject = Mcsv,
-                                params_rawdata = params_rawdata, params_general = params_general)
-  expect_equal(nrow(movisens_read$P$data), movisens_blocksize)
-  expect_false(movisens_read$filequality$filecorrupt)
-  expect_false(movisens_read$filequality$filetooshort)
-  expect_equal(sum(movisens_read$P$data[c("x","y","z")]), 4383.67, tolerance = .01, scale = 1)
-  expect_equal(movisens_read$endpage, movisens_blocksize)
-  
-  # read the next block (set PreviousEndPage to movisens_read$endpage)
-  movisens_read2 = g.readaccfile(movisensFile, blocksize = movisens_blocksize, blocknumber = 2, filequality = filequality,
-                                 dayborder = dayborder, ws = 3,
-                                 PreviousEndPage = movisens_read$endpage, inspectfileobject = Mcsv,
-                                 params_rawdata = params_rawdata, params_general = params_general)
-  expect_equal(nrow(movisens_read2$P$data), movisens_blocksize)
-  expect_equal(movisens_read2$endpage, movisens_blocksize * 2)
-  
-  # if the 1st sample of 2nd block is identical to the last sample of the 1st block,
-  # this means that we calculated the startpage of the 2nd block incorrectly.
-  expect_false(any(movisens_read2$P$data[1,] == movisens_read$P$data[nrow(movisens_read$P$data),]))
+  if (!Sys.getenv("CI") == "true") {
+    cat("\n Movisens")
+    
+    output_dir = "output_unisensExample"
+    on.exit({if (file.exists(output_dir)) unlink(output_dir, recursive = TRUE)}, add = TRUE)
+    if (file.exists(output_dir)) unlink(output_dir, recursive = TRUE)
+    
+    zip_file = "0.3.4.zip"
+    on.exit({if (file.exists(zip_file)) unlink(zip_file)}, add = TRUE)
+    if (!file.exists(zip_file)) {
+      # link to a tagged release of Unisens/unisensR github repo
+      movisens_url = "https://github.com/Unisens/unisensR/archive/refs/tags/0.3.4.zip"
+      download.file(url = movisens_url, destfile = zip_file, quiet = TRUE)
+    }
+    movisens_dir = "unisensR-0.3.4"
+    on.exit({if (file.exists(movisens_dir)) unlink(movisens_dir, recursive = TRUE)}, add = TRUE)
+    if (file.exists(movisens_dir)) {
+      unlink(movisens_dir, recursive = TRUE)
+    }
+    unzip(zipfile = zip_file, exdir = ".")
+    movisensFile = file.path(getwd(), "unisensR-0.3.4/tests/unisensExample/acc.bin")
+    
+    Mcsv = g.inspectfile(movisensFile, desiredtz = desiredtz)
+    expect_equal(Mcsv$monc, MONITOR$MOVISENS)
+    expect_equal(Mcsv$dformc, FORMAT$BIN)
+    expect_equal(Mcsv$sf, 64)
+    
+    movisens_blocksize = 3000
+    movisens_read = g.readaccfile(movisensFile, blocksize = movisens_blocksize, blocknumber = 1, filequality = filequality,
+                                  dayborder = dayborder, ws = 3,
+                                  PreviousEndPage = 1, inspectfileobject = Mcsv,
+                                  params_rawdata = params_rawdata, params_general = params_general)
+    expect_equal(nrow(movisens_read$P$data), movisens_blocksize)
+    expect_false(movisens_read$filequality$filecorrupt)
+    expect_false(movisens_read$filequality$filetooshort)
+    expect_equal(sum(movisens_read$P$data[c("x","y","z")]), 4383.67, tolerance = .01, scale = 1)
+    expect_equal(movisens_read$endpage, movisens_blocksize)
+    
+    # read the next block (set PreviousEndPage to movisens_read$endpage)
+    movisens_read2 = g.readaccfile(movisensFile, blocksize = movisens_blocksize, blocknumber = 2, filequality = filequality,
+                                   dayborder = dayborder, ws = 3,
+                                   PreviousEndPage = movisens_read$endpage, inspectfileobject = Mcsv,
+                                   params_rawdata = params_rawdata, params_general = params_general)
+    expect_equal(nrow(movisens_read2$P$data), movisens_blocksize)
+    expect_equal(movisens_read2$endpage, movisens_blocksize * 2)
+    
+    # if the 1st sample of 2nd block is identical to the last sample of the 1st block,
+    # this means that we calculated the startpage of the 2nd block incorrectly.
+    expect_false(any(movisens_read2$P$data[1,] == movisens_read$P$data[nrow(movisens_read$P$data),]))
+    
+  } else {
+    cat("Movisens testing skipped because depends on ability to download a file")
+  }
   cat("\n ad-hoc csv file")
   # ad-hoc csv file
